refactored teacher classes into #Teacher parrent class.
