+++ teacher_parrent.rb
@@ -1,0 +1,49 @@
+class Teacher
+  attr_reader :age, :salary, :phase
+  attr_accessor :name
+
+  def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @target_raise = nil
+    @phase = 3
+  end
+
+  def teach_stuff
+    response = ""
+    if self.class == ApprenticeTeacher
+      response += "Listen, class, this is how everything works. "
+      response += "*drops crazy knowledge bomb* "
+      response += "... You're welcome."
+    elsif self.class == SeniorTeacher
+      response += "Listen, class, this is how everything works, fo SHO! "
+      response += "*drops flat-out insane knowledge bomb* "
+      response += "... You're welcome. *saunters away*"
+    else
+      response += "*drops knowledge... on the floor* "
+    end
+    response
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if (rating > 80 && self.class == ApprenticeTeacher) || (rating > 90 && self.class == SeniorTeacher)
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+
+end
